@@ -1,50 +1,50 @@
 {
   "name": "accommodate",
-  "version": "0.0.0",
+  "productName": "Accommodate",
+  "version": "0.1.0",
   "main": "app.js",
   "private": true,
   "scripts": {
     "start": "electron .",
     "build": "rollup -c -w",
-    "serve": "run-p start build"
+    "serve": "run-p start build",
+    "distLinux": "electron-packager . accommodate --overwrite --asar=true --platform=linux --arch=x64 --icon=hexagon.png --prune=true --out=release-builds"
   },
   "dependencies": {
-    "cookie-parser": "^1.4.4",
+    "cookie-parser": "^1.4.5",
     "debug": "~2.6.9",
     "dotenv": "^8.2.0",
     "eases-jsnext": "^1.0.10",
-    "fuse.js": "^3.4.6",
+    "electron-packager": "^15.0.0",
+    "fuse.js": "^3.6.1",
     "morgan": "~1.9.1",
     "nedb": "^1.8.0",
-    "papaparse": "^5.1.1",
+    "papaparse": "^5.2.0",
     "parchment": "^1.1.4",
     "puppeteer": "^2.1.1",
-    "quill": "^2.0.0-dev.2",
-    "quill-better-table": "^1.2.8",
-    "quill-delta-to-html": "^0.11.0",
-    "svelte-calendar": "^1.3.4",
+    "svelte-calendar": "^1.5.0",
     "svelte-sortable-list": "^1.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.7",
-    "@babel/preset-env": "^7.7.7",
-    "autoprefixer": "^9.7.4",
-    "chokidar": "^3.3.1",
+    "@babel/core": "^7.11.1",
+    "@babel/preset-env": "^7.11.0",
+    "autoprefixer": "^9.8.6",
+    "chokidar": "^3.4.2",
     "cssnano": "^4.1.10",
-    "electron": "^7.1.9",
-    "node-sass": "^4.13.1",
+    "electron": "^7.3.2",
+    "node-sass": "^4.14.1",
     "nodemon": "^1.19.4",
     "npm-run-all": "^4.1.5",
     "postcss-preset-env": "^6.7.0",
-    "rollup": "^1.27.14",
-    "rollup-plugin-babel": "^4.3.3",
+    "rollup": "^1.32.1",
+    "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-commonjs": "^9.3.4",
-    "rollup-plugin-livereload": "^1.0.0",
+    "rollup-plugin-livereload": "^1.3.0",
     "rollup-plugin-node-resolve": "^4.2.3",
-    "rollup-plugin-postcss": "^2.0.6",
-    "rollup-plugin-svelte": "^5.1.1",
-    "rollup-plugin-terser": "^5.0.4",
-    "svelte": "^3.16.7",
-    "svelte-preprocess": "^3.4.0"
+    "rollup-plugin-postcss": "^2.9.0",
+    "rollup-plugin-svelte": "^5.2.3",
+    "rollup-plugin-terser": "^5.3.0",
+    "svelte": "^3.24.1",
+    "svelte-preprocess": "^3.9.12"
   }
 }
