@@ -24,7 +24,7 @@ function createWindow () {
     // and load the index.html of the app.
     win.loadFile('./index.html')
 
-    win.webContents.openDevTools()
+    // win.webContents.openDevTools()
 
     win.on('closed', ()=> {
         watcher.close()
